+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## [UNRELEASED] prometheus-gitlabci-exporter 0.1.1
+## [UNRELEASED] prometheus-gitlabci-exporter 0.2.0
+
+### Features
+- add GitLab host name metrics ([#4](https://github.com/AleksanderWWW/prometheus-gitlabci-exporter/pull/4))
 
 ### Changes
 - refactor to make code more testable ([#3](https://github.com/AleksanderWWW/prometheus-gitlabci-exporter/pull/3))